fix : issue with finding main.py
